Restructured file's architecture and added some comments
