+++ spell-checker.cabal
@@ -31,19 +31,20 @@ library spell-checker-lib
 
 executable spell-checker
     main-is:    Main.hs
+    other-modules: SpellCheckerInterface
 
     build-depends:
         spell-checker:spell-checker-lib,
         base        ^>= 4.19.0.0,
-        vector      ^>= 0.13.1.0,
         containers  ^>= 0.6.8,
         directory   ^>= 1.3.8.3,
         mtl         ^>= 2.3.1,
-        bytestring  ^>= 0.12.1.0,
         serialise   ^>= 0.2.6.1,
-        no-buffering-workaround ^>= 0.1.0.0,
+        vector      ^>= 0.13.1.0,
         aeson       ^>= 2.2.1.0,
-        text        ^>= 2.1.1
+        bytestring  ^>= 0.12.1.0,
+        text        ^>= 2.1.1,
+        no-buffering-workaround ^>= 0.1.0.0
 
     hs-source-dirs:   src
     default-language: Haskell2010
