+++ src/Main.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE OverloadedStrings, DeriveFunctor #-}
+{-# LANGUAGE DeriveFunctor #-}
 {-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
 
 -- --------------------------------------------------------------------------
@@ -20,15 +20,6 @@ import Control.Monad.IO.Class (liftIO)
 import Control.Monad (when)
 
 import Data.Char (isSpace)
-import Data.List (sort, sortOn)
-import Data.Maybe (mapMaybe, fromJust)
-import Data.Ord (Down(Down))
-import Data.Vector (Vector, imap, (!), (!?))
-import Data.Aeson (decode, Array, Value(Object, Array, String))
-import qualified Data.Aeson.KeyMap as KM ((!?), Key)
-import qualified Data.Vector as V (find, map, fromList, zip, zipWith, concat, toList)
-import qualified Data.ByteString.Lazy as B (readFile)
-import qualified Data.Text as T (head)
 
 import System.IO (IOMode (ReadMode), hSetBuffering,
     hFlush, stdout, stdin, BufferMode (NoBuffering))
@@ -36,22 +27,8 @@ import System.IO.NoBufferingWorkaround (initGetCharNoBuffering, getCharNoBufferi
 import System.Info (os)
 import Codec.Serialise (readFileDeserialise)
 
-import Data.WordTree
-
--- | Used Keyboard, the only one supported for now is QWERTY
-type Keyboard = Vector (Vector Char)
-
--- | Entrypoint into the spell-checker
-main :: IO ()
-main = do
-  if os == "mingw32" then initGetCharNoBuffering >> hSetBuffering stdout NoBuffering
-  else hSetBuffering stdin NoBuffering
-  inputFreq <- readFileDeserialise "SerializedStatistics/result" 
-  let dictionaryTree = fromMap inputFreq
-  putStrLn "Type enter to correct a word or tab to complete it."
-  putStrLn "Type a word:"
-  (`runContT` return) $ do
-    callCC $ prompt dictionaryTree
+import Data.WordTree (Tree, fromMap)
+import SpellCheckerInterface (completeWord, correctWord)
 
 -- | The user can either ask for completion or correction
 data Query a = Complete a | Correct a deriving Functor
@@ -82,88 +59,15 @@ prompt tree exit = do
     Correct word -> liftIO $ print =<< correctWord tree word
   restart
 
--- | Complete user's input
-completeWord :: Tree Char -> String -> [CountedWord]
-completeWord tree prefixe = take 10 . sortOn Down $ giveSuffixe tree prefixe
+-- | Entrypoint into the spell-checker
+main :: IO ()
+main = do
+  if os == "mingw32" then initGetCharNoBuffering >> hSetBuffering stdout NoBuffering
+  else hSetBuffering stdin NoBuffering
+  inputFreq <- readFileDeserialise "SerializedStatistics/result" 
+  let dictionaryTree = fromMap inputFreq
+  putStrLn "Type enter to correct a word or tab to complete it."
+  putStrLn "Type a word:"
+  (`runContT` return) $ do
+    callCC $ prompt dictionaryTree
 
-associateDistance :: String -> CountedWord -> IO (Int, CountedWord)
-associateDistance w x = do
-    distance <- strDiff w (word x)
-    return (distance, x)
-
--- | Correct user's input
-correctWord :: Tree Char -> String -> IO [CountedWord]
-correctWord tree word = do
-    wordsWithDistance <- mapM (associateDistance word) $ similarWords tree 2 word
-    return $ take 10 . map snd $ sortOn fst wordsWithDistance
-
--- | [Unused for now] Correct a whole line, giving only one answer
--- correctLine :: Tree Char -> String -> CountedWord
--- correctLine tree line = assemble correctWords
---   where
---     correctWords = map (head <$> correctWord tree) $ words line
---     assemble = foldl (\w1 w2 -> CountedWord (show w1 ++ " " ++ show w2) (WordProperties (negate 1) []))
---                       (CountedWord "" nullProperties)
-
--- | Extracts the text value out of a json
-getFromJSON :: KM.Key -> Value -> Char
-getFromJSON key (Object o) = (\(String s) -> T.head s) . fromJust $ o KM.!? key
-
--- | QWERTY Keyboard used to get neighboors leters from the on typed 
-keyboardEn :: IO Keyboard
-keyboardEn = do
-  layoutB <- B.readFile "layouts/qwerty.json"
-  let layoutArray = fromJust (decode layoutB :: Maybe Array)
-  let fromJSON = V.map (\(Array v) -> V.map (getFromJSON "label") v) layoutArray
-  return fromJSON
-
--- | Gives the zone of index of near characters from the index of the one given
-nearIndices :: (Num a, Num b, Eq a, Eq b) => a -> b -> [(a, b)]
-nearIndices indR indT = [(row, text) | row <- [indR - 1, indR, indR + 1],
-                                        text <- [indT - 1, indT, indT + 1],
-                                        (row, text) /= (indR, indT)]
-
-accessMatrix :: Keyboard -> (Int, Int) -> Maybe Char
-accessMatrix matrix (row, column) = (!? column) =<< (matrix !? row)
-
--- | Give the matrix of neighboorhood in place of the character of the keyboard 
-charsPerimeter :: Keyboard -> Vector (Vector [Char])
-charsPerimeter keyboard = imap (\indR row -> imap
-                                              (\indT _ -> mapMaybe (accessMatrix keyboard) $ nearIndices indR indT)
-                                              row)
-                                keyboard
-
--- | Given a keyboard and a perimeter, associate between each characters and his neighboors
-associateNearChars :: Keyboard -> Vector (Vector [Char]) -> Vector (Char, [Char])
-associateNearChars keyboard perimeter = V.concat . V.toList $ V.zipWith V.zip keyboard perimeter
-
--- | Gives the neighboors of all characters
-nearChars :: Keyboard -> Vector (Char, [Char])
-nearChars keyboard = associateNearChars keyboard $ charsPerimeter keyboard
-
--- | The keyboard we choose to use
-actualKeyboard :: IO (Vector (Char, [Char]))
-actualKeyboard = nearChars <$> keyboardEn
-
-inPerimeterOf :: Char -> IO String
-inPerimeterOf c = maybe [] snd . V.find ((c == ) . fst) <$> actualKeyboard
-
-min' :: IO Int -> IO Int -> IO Int
-min' a b = min <$> a <*> b
-
--- | Calcul of the distance between two words (modifed Hamming's distance)
-strDiff :: String -> String -> IO Int
-strDiff x [] = return $ length x
-strDiff [] y = return $ length y
-strDiff (x : xs) (y : ys) =
-    if x == y then
-      strDiff xs ys
-    else do
-      perimeterOfy <- inPerimeterOf y
-      (+2) <$> diffMinq perimeterOfy
-  where
-    diffMin perimeterOfy = min' (diffMinq perimeterOfy) $ min' (strDiff xs (y:ys)) (strDiff (x:xs) ys)
-    -- Case where they are "near" :
-    diffMinq perimeterOfy = do
-        distance <- strDiff xs ys
-        return $ distance - fromEnum (x `elem` perimeterOfy)
